+++ Engine/script/script_runtime.h
@@ -59,4 +59,4 @@ extern void ccSetDebugHook(new_line_hook_type jibble);
 extern void ccSetScriptAliveTimer (int);
 // reset the current while loop counter
 extern void ccNotifyScriptStillAlive ();
-extern int call_function(long addr, int numparm, const RuntimeScriptValue *parms, int offset);
+extern int call_function(intptr_t addr, int numparm, const RuntimeScriptValue *parms, int offset);
